Feat: ReviewImgDto 메서드 분리, Input/Output Dto 생성, 예외처리, 유효성검사

+++ src/main/java/cyy/greenblue/repository/ReviewImgRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ReviewImgRepository extends JpaRepository<ReviewImg, Integer> {
+public interface ReviewImgRepository extends JpaRepository<ReviewImg, Long> {
     List<ReviewImg> findByReview(Review review);
 }
